feat(rn): 相关逻辑迁移至 editTemplate.ts 文件
